NoteScape
📜 NoteScape 2.0 License and Usage Information

NoteScape 2.0 is developed and licensed by Nutreon Studio. By continuing to use this software, you agree to the following license terms.

Purpose of Use: The software is intended for personal and non-commercial use only.

Intellectual Property: All rights belong to Nutreon Studio (© 2025). The software, in whole or in part, may not be copied, distributed, or modified without permission.

Disclaimer of Liability: The software is provided “as is.” Nutreon Studio is not responsible for any direct or indirect damages resulting from the use of the software.

Legal Actions: In case of any violation of the license terms, Nutreon Studio reserves the right to take legal action.

For any questions or support requests, please contact:
✉ tirasmustafa033@gmail.com

Thank you for choosing NoteScape 2.0. We wish you productive use.

The Nutreon Studio Team